+++ src/main/java/aresain/loldatastats/lolstat/statanalysis/StatAnalysisService.java
@@ -17,9 +17,9 @@ import aresain.loldatastats.entity.ParticipantSummary;
 import aresain.loldatastats.loldata.gamematch.GameMatchService;
 import aresain.loldatastats.loldata.gamematch.dto.GameMatchInfoDto;
 import aresain.loldatastats.lolstat.statanalysis.accumulator.SummaryAccumulator;
-import aresain.loldatastats.lolstat.statanalysis.dto.AnalysisListDto;
-import aresain.loldatastats.lolstat.statanalysis.dto.ObjectiveAnalysisDto;
-import aresain.loldatastats.lolstat.statanalysis.dto.SummaryAnalysisDto;
+import aresain.loldatastats.lolstat.statanalysis.dto.match.AnalysisListDto;
+import aresain.loldatastats.lolstat.statanalysis.dto.match.ObjectiveAnalysisDto;
+import aresain.loldatastats.lolstat.statanalysis.dto.match.SummaryAnalysisDto;
 import aresain.loldatastats.lolstat.statanalysis.repository.ObjectiveStatRepository;
 import aresain.loldatastats.lolstat.statanalysis.repository.ParticipantSummaryStatRepository;
 import lombok.RequiredArgsConstructor;
